% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.xgb.R
\name{mice.impute.xgb}
\alias{mice.impute.xgb}
\title{Description of your function.}
\usage{
mice.impute.xgb(
  y,
  ry,
  x,
  wy = NULL,
  xgb.params = NULL,
  nthreads = -1,
  match.type = "predicted",
  k = 3,
  ...
)
}
\arguments{
\item{y:}{Vector of values of response variable}

\item{ry:}{Vector of true/false representing observed values in target variable. TRUE representing respective value was observed and FALSE represents missing value.}

\item{wy:}{Vector of true/false representing missing values in target variable. TRUE representing respective value was missing and FALSE represents observed value.}

\item{x:}{dataframe. Data of all variables except target variable.}

\item{xgb.params:}{list of XGBoost parameters obtained from xgb_param_calc function. User can also manually provide a list of parameters.}

\item{nthreads:}{Number of threads. By default, function uses all available - 1 cores.}

\item{match.type:}{Possible value can be "predicted.observed", "original.observed" or "predicted"}
}
\value{
Imputed values for target values and training error (actual - predicted)

XGBoost uses a set of hyperparameters for tuning the algorithm. User has three options to select these parameters
1) Using default parameters (adapted from Deng et al., 2023).
2) Run xgb_param_calc function to let function randomly select one variable as a target and tune the parameters using only one target variable.
3) Run xgb_param_calc function to iteratively select best set of hyperparameters for all variables.

mice.impute.xgb
}
\description{
Description of your function.
}
\examples{
mice(nhanes, method = "xgb", param = NULL)

}
