% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_error.R
\name{testing_error}
\alias{testing_error}
\title{Calculation of testing_error. This function can be used on the output of mice.impute.xgb.}
\usage{
testing_error(orig_data_1, imputed_res_1)
}
\arguments{
\item{orig_data_1:}{original data without missing values}

\item{imputed_res_1:}{imputation results from mice.impute.xgb}
}
\value{
test error list of all variables in the dataset. Each list contain m values for m different imputations

testing_error
}
\description{
Calculation of testing_error. This function can be used on the output of mice.impute.xgb.
}
