% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_param_calc.R
\name{xgb_param_calc}
\alias{xgb_param_calc}
\title{Creates a \code{xgb_param_calc} argument}
\usage{
xgb_param_calc(
  data,
  response = NULL,
  select_features = NULL,
  num_cores = 1,
  iter = 30
)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.  Missing
values are coded as \code{NA}.}
}
\value{
Character vector of \code{ncol(data)} element
}
\description{
This helper function creates a valid \code{xgb_param_calc} vector. The
\code{xgb_param_calc} vector is an argument to the \code{mice} function that
specifies post-processing for a variable after each iteration of imputation.
}
\examples{
make.post(nhanes2)
}
\seealso{
\code{\link{mice}}
}
